Changed pagination count to int.

+++ src/main/java/com/after_sunrise/cryptocurrency/bitflyer4j/entity/CoinIn.java
@@ -37,7 +37,7 @@ public interface CoinIn {
     class Request extends Entity implements Pagination {
 
         @SerializedName("count")
-        private final Long count;
+        private final Integer count;
 
         @SerializedName("before")
         private final Long before;
