+++ src/main/java/com/after_sunrise/cryptocurrency/bitflyer4j/entity/CoinOut.java
@@ -41,7 +41,7 @@ public interface CoinOut {
     class Request extends Entity implements Pagination {
 
         @SerializedName("count")
-        private final Long count;
+        private final Integer count;
 
         @SerializedName("before")
         private final Long before;
