+++ src/main/java/com/after_sunrise/cryptocurrency/bitflyer4j/entity/Execution.java
@@ -39,7 +39,7 @@ public interface Execution {
         private final String product;
 
         @SerializedName("count")
-        private final Long count;
+        private final Integer count;
 
         @SerializedName("before")
         private final Long before;
