+++ src/main/java/com/after_sunrise/cryptocurrency/bitflyer4j/entity/TradeCollateral.java
@@ -33,7 +33,7 @@ public interface TradeCollateral {
     class Request extends Entity implements Pagination {
 
         @SerializedName("count")
-        private final Long count;
+        private final Integer count;
 
         @SerializedName("before")
         private final Long before;
